Works with moving obstacles
Hook and Rope work with moving obstacles.
The hook is now a child of whatever obstacle it lands on and scales its local scale to maintain a fixed size.
Need to change hook flying mechanism otherwise it will change direction to hit a predetermined location on the obstacle.
The moving obstacles currently make the rope inconsistent in length