Big Changes to Rope Mechanism
No longer use "corner colliders" as decided they produced too many problems.
Rope splitting now works by determining if an object is hit by a raycast to the child. If it is then the rope will split at the nearest vertex of the collider. Polygon2D colliders must now be used for all colliders to enable this functionality.
The position of the nodes must now update to the vertex of the collider which it attaches too which is stored as a variable. A function specifically for updating the position has been added as this became more complicated as a result.
Recombination of rope components now works through determining if the object the child is attached to overlaps with a triangular region defined by the node, its child and its grandchild. This is a seemingly foolproof approach. A few functions associated with determining if arbitrary polygons are overlapping have been added to enable this.