Improved Split Detection
Splitting point is now determined by the vertex whos distance is closest to the line at the time of collision. The vertex of the current node and its child are not included when finding the nearest vertex.
Two functions have been added to determine the distance of a point to a line and to find which point in an array of points is closest to a line (with the possibility of excluding a vertex in the array by index)
These changes have made the splitting mechanism far more robust. It is now very difficult to break it.

The render function has been put in late update. This prevents the line from rendering when the things the nodes are attached to have been updated but the rope nodes themselves havent.