Hook Attaching Improved
Point where hook latches no longer determined at the point of firing. The hook now travels as a collider until it reaches an obstacle collider or its LoS to the ball is blocked (then it will destroy itself). If hitting a obstacle collider it will attach to the surface and move forward or backward along the travel direction a distance such that it is a predetermined separation distance from the collider surface (assumes the collider is flat in region it has had to move)
The hook also no longer becomes the child of the object it attaches to. It instead stores what obstacle it collided with and the relative position it has to the obstacle in the obstacles local coordinates. The position of the hook is then updated every frame to maintain that relative position.