ColorPalette changes and small change to rope dynamic
ColorPalettes changed so now I have a number of prefab object that represent color palettes. Makes them easier to manage and swap in and out.

Clicking on the screen now makes the hook fire if there is not one already, or cancels any existing hook.

Now the final node of the rope adjusts its max length to be the current separation to the player. This makes the ball flow better and makes it easier to maintain speed. Essentially the rope is always taught but you can move towards the node.