Menu systems + improved color system + other bits
ColorPalettes are now ScriptableObjects and are easier to use. Objects with a ColorManager can set which index of the color palette they wish to use.

BallHandler improved to ensure velocity of ball before and after detaching rope is the same.

GameManager methods are now all static, meaning they are callable from anywhere without needing a reference to an actual object. They are still attached to objects current to store what screen orientation a scene has though this will be changed later.

Menu systems added, generally the buttons on the menus screens do something and then run a GameManager function.

MovingPlatform improved to give local coordinates as start and end point. Will change at some point to be able to next change to have a list of destinations and whether the object just ends at the final one, goes back in reverse, or loops back to the first position (will do this when i actually want something to do that).