Checkpoints + small temporary fixes
Checkpoints added, function as trigger collider which updates the players respawn position to their own position.

Made some small temporary fixes as quick solutions for togglable wall of death and the trailrenderer of the ball when respawning. Did this quick as I go on holiday tomorrow and would like to have a functional version to show people for some feedback.