Reverted to fixed max rope length
Rope now remembers its max length instead of shortening to the current separation as it was not intuitive enough even if it felt better some of the time.
Also just cleaned up the MaintainLength function as was a little difficult to understand exact;y what was happening.
Also added the quality options setting after discovering it did not run smoothly enough on a cheap, low end phone.